Make sure to log execv errors from "bazel run".
This makes all helper Execute* functions "final" by marking them
ATTRIBUTE_NORETURN and ensures that the auxiliary ExecuteProgram is the
one to log execv errors.

I tried writing a test for this but it seems pretty hard: "bazel run"
uses a helper script to run the test program under, so tricking it to
run an invalid binary is not sufficient.

RELNOTES: None.
PiperOrigin-RevId: 270738286