Rewrite documentation on rule outputs
Separates that out into a subsection of the section on Files.

Start with the preferred way to create and propagate predeclared and non-predeclared outputs. That avoids the parts about the preferred way to do things being split up by a long section on the deprecated way to do things, and it eliminates a bit of repetition.

For each deprecated way of adding predeclared outputs, add a little bit about what to do instead. For the bit about the outputs parameter of the rule function, go a little less in detail about the syntax, since it already links to the documentation. For the bit about ctx.outputs.executable, reword a little to make it clearer that only use of ctx.outputs.executable is deprecated, not the "executable" and "test" parameters of the rule function.

Some technical details that are more confusing than useful are just omitted:

* A parenthetical remark about how the default executable output is the one predeclared output that doesn't have a label is pretty confusing, but generally doesn't matter to buildrule authors.

* A section detailing the behavior of ctx.attr and ctx.file for output attrs seemed like an unnecessarily long digression. Better to just say "use ctx.outputs".

RELNOTES: Improve documentation on rule outputs.
PiperOrigin-RevId: 270928558