bazel syntax: delete SkylarkImport
https://github.com/bazelbuild/bazel/commit/b2e1fe367269103e127718131163cd702cefcdff caused skyframe to check load strings for validity,
so the parse-time checks are now redundant; this change deletes them.

The skyframe checks require only that the strings are valid labels,
so load("foo.bzl") is equivalent to load(":foo.bzl").
Before, the colon was required, as a consequence of the obsolete
strategy used to transition away from non-label filenames.
The treatment of label strings by load is now consistent with
other label strings in a BUILD file.

This change removes another lib.syntax->lib.cmdline dependency.
The remaining two are GlobalFrame.label (broken in unknown commit)
and GlobalFrame.callerLabel, used for the Starlark Label function
(broken in unknown commit).

Also:
- PackageFunctionTest: add tests of load validation.
- ParserTest: break (subclass) dependency on EvaluationTestCase
- BuildFileAST: remove all mention of RepositoryMapping.
- SkylarkEvaluationTests: delete obsolete tests.
- ValidationTest: move tests for toplevel for/if statements here.
- ParserTest: delete load validation checks.
- ParserInput: remove spurious IOException and simplify one caller.
PiperOrigin-RevId: 270958511