bazel cmdline: make cmdline depend on syntax, a reversal
Historically, the interpreter had many dependencies on Label.
Now, they are all broken, and the syntax->cmdline edge can be broken.
This change does that, and establishes the reverse dependency,
allowing Label.relative to use Environment instead of StarlarkContext,
which is scheduled for deletion (supplanted by Environment.getThreadLocal)
in a later change.

PiperOrigin-RevId: 270969911