bazel syntax: abolish Environment.StarlarkContext
The bulk of this change eliminates parameters of type
StarlarkContext in annotated methods, replacing them
by Environment parameters if not already present.
From the Environment, the BazelStarlarkContext can
be retrieved using getThreadLocal.

The only interesting parts are:

Environment: delete {get,set}StarlarkContext;
use {get,set}ThreadLocal instead.
Also, sprinkle 'private' keywords.

Label: store the HasRepoMapping as a thread-local.
Its value is an alias for the BazelStarlarkContext, but
we can't mention that type down here.

BazelStarlarkContext: define and use BSC.from(Environment),
which loads the Bazel info from a Starlark thread,
and BSC.storeInThread(Environment), which stores it,
plus the Label.RepoMapping, so we don't forget it.

ConfiguredRuleClassProvider: inline one flavor of
createSkylarkRuleClassEnvironment into the other.
PiperOrigin-RevId: 271039685