Fix crash in CollectLocalCpuUsage.
It is possible for time to travel "backwards" (currentTimeMillis() returning a value smaller than an earlier call to the same method). If time traveled backwards by at least 2000 milliseconds then this code would try to create an array with a negative length, causing an exception.

This change doesn't fix the underlying problem (the class performs a lot of unsafe time math) but bounds the array creation to not use negative values.

RELNOTES: None
PiperOrigin-RevId: 271572798