Catch strategies that call StopConcurrentSpawn more than once.
This (and other debug code I added but then removed because it should
not be part of this) shines light on why I was having trouble with
using a Semaphore in stopBranch().

So instead of using a Sempahore, use an AtomicReference (like the
legacy spawn scheduler does) to track the strategy that was able to
run the stopBranch code. This also avoid having to worry about
the strategies doing the right thing regarding the interrupt bit
status. (I'm now more convinced that they do the right thing.)

RELNOTES: None.
PiperOrigin-RevId: 272283111