Add explicit plumbing for a "validation artifact" to AndroidResourcesInfo
This unblocks fixing a bug in which the output of validation actions [1] is
used for creating the top-level [2] resources.jar file.  Instead, the actions
which generate level-library [3] resource JARs should dictate the contents of
the top-level resources.jar

The extra plumbing is needed to allow running the validation actions in parallel
with the rest of the build if their outputs are not needed for any other purpose.

[1] https://github.com/bazelbuild/bazel/blob/1664d54293749a58d4dbdcf06c9c203de84f1ab1/src/main/java/com/google/devtools/build/lib/rules/android/AndroidResourceValidatorActionBuilder.java#L154
[2] https://github.com/bazelbuild/bazel/blob/1664d54293749a58d4dbdcf06c9c203de84f1ab1/src/main/java/com/google/devtools/build/lib/rules/android/RClassGeneratorActionBuilder.java#L89
[3] https://github.com/bazelbuild/bazel/blob/1664d54293749a58d4dbdcf06c9c203de84f1ab1/src/main/java/com/google/devtools/build/lib/rules/android/AndroidResourceMergingActionBuilder.java#L201

PiperOrigin-RevId: 272341623