Extract a CommonPrerequisiteValidator class
The new class takes over most of the responsibilities of
BazelPrerequisiteValidator, except that it delegates to two abstract
methods for definitions of whether two packages are logically equivalent
(e.g., packages under a source tree and the corresponding test tree),
and whether a package is considered experimental.

This adds a new conceptual check that code outside of experimental may
not depend on code inside experimental. However, Bazel does not provide
a policy for what is considered experimental, so there is effectively no
enforcement at this time.

This consistutes a small semantic change: previously, Bazel used
different definitions for 'same logical package' for visibility, which
used 'equal package identifiers', and deprecation, which used
'java/javatests are the same logical package'. Now, Bazel uses 'equal
package identifiers' for both. Unfortunately, this could increase the
number of deprecation warnings for users who rely on this, but it does
not cause anything to break.

In the future, we may add a mechanism to configure these. For now, this
improves code sharing between open and closed source code.

PiperOrigin-RevId: 272402840