Describe known issue when in-mem state is lost when running bazel builds inside docker containers e.g. in CI
I hit that problem and discovered answer by reading that doc: https://docs.bazel.build/versions/master/user-manual.html#flag--batch

I was trying to understand why my jobs on CI takes around 20s even with caching but in dockers when locally it was like 200 ms. So maybe it would be better to place it in docs for future generations.

Closes #6657.

PiperOrigin-RevId: 272406338