Stop lexer from scanning characters in ints if first token in not a zero
If the first token of a number is not '0', then the lexer should not scan any characters because the number should be a decimal.
This solves some cases where omitting whitespace between tokens produced errors. For example
```
# Before
6in() # False
6or() # invalid base-10 integer constant: 6o

# After
6in() # False
6or() # 6
```
The first example works correctly in both cases since the token 'i' never advances the buffer, while 'o' can appear in integers.

Closes #9210.

PiperOrigin-RevId: 272431243