JavacOptions: Fix incompatibility of -target and -parameters options
Closes #8772.

Java compiler -target 7 (and lower) and -parameters options are mutually
exclusive. Starting from version 11 the javac is issuing this warning:

  warning: -parameters is not supported for target value 1.7. Use 1.8 or later.

and is spamming the build log. This is particularly annoying, because
the major project in distributed Java ecosystem Google Protobuf is
targeting java language level 7, so that this spam affects everyone in
the wild. But the scope of the problem is unrelated to Protobuf and
every attempt to pass -target 7 javac option to the java_library rule
ends up in the above spam.

To rectify extend ReleaseOptionNormalizer and filter out -parameters
option from the command line if java language level is 7 or lower.

Test Plan:

* Build new java_tools with:

  $ bazel build //src:java_tools_java11.zip

* Switch to using custom version of java_tools, e.g.:

  http_archive(
    name = "local_java_tools",
    urls = ["file:///tmp/java_tools_java11.zip"],
  )

* Build java library with custom java tools toolchain:

  $ cat A.java:
class A {}

  $ cat BUILD:
java_library(
    name = "a",
    srcs = ["A.java"],
    javacopts = [
        "-source 7",
        "-target 7",
    ],
)

  $ bazel build \
    --host_java_toolchain=@local_java_tools//:toolchain \
    --java_toolchain=@local_java_tools//:toolchain \
    :a

and confirm that there is no build warnings.

* Retry the build without custom java tools and confirm that the warning
is reported:

  $ bazel build :a
  warning: -parameters is not supported for target value 1.7. Use 1.8 or later.

Change-Id: Ibec6dd14d8d4a395a6de3128cd15ccec52752d4b