bazel syntax: remove type parameters from Parameter<V, T>
Parameter is a part of the syntax, and is now created only by
the parser. Its "default value" (V) is always an Expression.
Its "type" (T) is no longer needed.

The type parameters were presumably added to achieve an unholy
form of code re-use in SkylarkSignatureProcessor and FunctionSignature:
the @Param annotations from Java methods were converted into
pseudo-syntax and then fed to WithValues.of.

This change forks rather than reuses WithValues.of.
One copy is renamed fromParameters, and is specialized to the parser;
the other copy is inlined into getSignatureForCallable and specialized
to no longer need the construction of transient Parameter syntax nodes.
Both copies will be simplified in follow-up changes that
make further slices at the monster that is FunctionSignature.

Also: remove confusingly named isMandatory and isOptional methods.
They were instanceof tests, and thus not opposites.
PiperOrigin-RevId: 272868319