bazel syntax: delete FunctionSignature.WithValues
WithValues was a triple of (signature, default values, types). This
change replaces it by three separate parameters, though in most cases
only one or two are needed.

FunctionSignature is now part of the :frontend portion of lib.syntax,
and thus cannot depend on the representation of values. Printing of
default values and parameter types is now parameterized by a
function passed in by the caller.

Also, reduce the scope of many unnecessarily public declarations.

With this change, the syntax trees no longer depend on the representation of values.
(See unknown commit)

PiperOrigin-RevId: 273337445