external.md: clarify where external repositories are fetched to
Document that external repositories are handled on a per-workspace
basis and in the section on "fetching" link to the section explaining
the directory layout. Also, be more explicit about what symlinks
are created; symlinks are a way to make local repositories stay up
to date and are not to be confused with the convenience symlinks
that bazel creates.

Fixes #8091

Change-Id: I8b35d03abb5f4f8a1b429ee2cb875ec71060b9db
PiperOrigin-RevId: 273533845