WindowsFileSystem.readSymbolicLink: Map IOException to NotASymlinkException
Fixes symbolic link outputs on Windows.

FileSystem.resolveOneLink expects that readSymbolicLink throws NotASymlinkException when the provided path is valid but is not a symbol link. UnixFileSystem.readSymbolicLink throws the exception when necessary, but WindowsFileSystem.readSymbolicLink - doesn't (throws IOException instead).

The PR fixes WindowsFileSystem.readSymbolicLink. It's implemented in the same way as Unix implementation.

Closes #9259.

PiperOrigin-RevId: 273715652