bazel syntax: hide EvalUtils.execToplevelStatement
All file execution now happens through EvalUtils.exec(StarlarkFile)
or EvalUtils.execOrEval (which I plan to eliminate in a followup).
Clients do not execute individual statements.

This requires a temporary hack: StarlarkThread.postAssignHook is a
hook that clients can install to be notified of each assignment at top
level. SkylarkImportLookupFunction uses it to "export" values such as
rules, aspects, and transitions during execution, instead of at the
end of execution. (The behavior will be changed soon; see unknown commit
and b/65374671.)

This may seem like two steps forward and one back, but it will allow us
to develop the EvalUtils API, in particular, turning the top-level statements
of a StarlarkFile into a StarlarkFunction, leading to a consistent treatment
of the call stack.

Also:
- EvalUtils.debugExec is now expressible in terms of the API,
  and has been moved up into the debugger.
- Identifier.boundIdentifiers is no longer public.
- Minor simplifications to tests.
PiperOrigin-RevId: 273741627