Don't filter out empty package(s) from the resource table during resource shrinking.
In the final copy stage following a shrink, if the app package (package id 7F) is empty, current logic will exclude the package from the APK, which will cause the subsequent invocation of `aapt2 link` to fail with a message "static library has no package".

It's insignificant to exclude empty packages in this context, as there would never be many.

PiperOrigin-RevId: 273795275