Add an option to allow local actions to run without locking outputs.
This can be enabled by setting --experimental_local_lockfree_output,
which in turn requires --legacy_spawn_scheduler=false.

When using the new spawn scheduler, the local spawn is actively
interrupted once the remote spawn completes and we are ready to download
its outputs. Because the interruption happens actively and at a specific
place, we can allow the local spawns to make progress concurrently with
the remote spawn. This is useful because, if there is a remote cache hit
soon, we want to make use of it even if the local spawn is already
running.

Note that this feature is still highly experimental and there is no
guarantee that it will work fine just yet. This is why I'm leaving the
option in the undocumented category, for example.

RELNOTES: None.
PiperOrigin-RevId: 273818854