Test result processing: hide TestResultAnalyzer
This is in preparation for refactoring the combination of
TestResultAnalyzer and AggregatingTestListener. At this time, they are
tightly coupled with both calling into each other. I think there's a
better structure hidden that allows us to separate them into
well-defined components, and also reduces the contention in
AggregatingTestListener which uses a single large lock for all test
processing; instead, we want to use a lock per test target.

This change also simplifies the callers which is nice.

This is also in preparation for allowing runs_per_test in combination
with runs_per_test_detects_flakes to interrupt concurrent attempts when
the first one comes back successfully, which should return end-to-end
latency for large flaky tests.

PiperOrigin-RevId: 273958739