Move runs_per_test_detects_flakes to the test config
This is in preparation for allowing successful test actions to cancel
parallel runs to reduce end-to-end latency for flaky tests. In order to
do this, the test action creation machinery needs to know if it's
running in this mode.

This could increase end-to-end latency by increasing the likelihood of
re-analysis (due to the flag now being a configuration option). However,
I can't think of any realistic scenario where this would happen: it
would require --runs_per_test already being set, and
--runs_per_test_detects_flakes being toggled, and no other configuration
flag changing. Maybe someone fixed a flaky test and now disables the
flag in order to check if it works, but with --runs_per_test already
set? Seems very far-fetched.

Also note that --runs_per_test currently disables caching for test
actions, so its runtime is likely dominated by execution rather than
analysis.

PiperOrigin-RevId: 274566875