Add support for preventing idle sleep on hardware/OS combos that support it.
RELNOTES: Blaze will prevent idle sleep during test and build actions. Note that this does not affect screen savers and will not keep a laptop awake if the user forces sleep or closes the lid. This is purely to avoid idle sleeping when the user is not interacting with the device.

Currently we support Windows and macOS.

PiperOrigin-RevId: 274585861