bazel syntax: use EvalUtils.{exec,eval} instead of execOrEval
execOrEval, now renamed execWithOptionalFinalExpression, accepts a
sequence of statements optionally followed by an expression. It is
appropriate for use in an interactive UI such as the REPL or debugger,
but should not be used elsewhere.

This change makes all non-REPL callers choose either to execute statements
for their effects, or to evaluate an expression for its value.

PiperOrigin-RevId: 274594764