Adds a special "_validation" output group to enable moving "validation actions" off the critical path of builds.
Validation actions are actions which validate something about the build based on execution-time information (e.g. the contents of files), but whose outputs are not needed by anything else in the build. Currently, to get such an action to run, its output (even a dummy empty file) must be "artificially" added to the inputs of some other action in the build. This reduces parallelism and potentially increases the critical path of the build.

The validation output group is provided here instead, so that the outputs of validation actions can be requested separately from all other targets, allowing the actions to be executed in parallel with the rest of the build.

The "_validation" output group is "special" in that it is automatically propagated through implicit outputs and through all rules. It also cannot be "deselected" with the --output_groups flag. Instead, the flag --run_validations is provided.

Note that with validation actions supported through output groups, such actions must still create some output file, even if it's empty.

RELNOTES[NEW]: Added a special "_validation" output group to enable moving "validation actions" off the critical path of builds.

PiperOrigin-RevId: 274652209