bazel syntax: rename StarlarkThread.GlobalFrame to Module
In the interests of expedient review, other than edits to the
Module doc comment, and adding package-visibility to Module's fields,
this is a pure renaming.

Many existing function and variable names are less than ideal; they
will be addressed in later changes.

This is a breaking API change for copybara.

PiperOrigin-RevId: 274859481