Fix a bug that can produce incorrect Starlark transitions.
This specifically affects incoming edge transitions that read attributes:
 - https://docs.bazel.build/versions/master/skylark/config.html#incoming-edge-transitions
 - https://docs.bazel.build/versions/master/skylark/config.html#accessing-attributes-with-transitions

And even that's not sufficient to trigger the bug. It also depends on the
order of a rule's dependencies and which transition attaches to which dep.

The basic problem is in ConfigurationResolver.resolveConfigurations, which is tasked
with figuring out the configuration for every dep of a rule (which implicitly requires
applying transitions). Specifically these lines: https://github.com/bazelbuild/bazel/blob/77497817b011f298b7f3a1138b08ba6a962b24b8/src/main/java/com/google/devtools/build/lib/analysis/config/ConfigurationResolver.java#L244-L246

These lines use a cache to save from having to compute transitions redundantly. e.g.
imagine a rule has 100 deps all of which apply a transition that sets foo=bar. All 100
of these deps use the same configuration. So there's no need to re-evaluate the
transition and re-request the dep configuration 100 times (and yes, doing so can
observably harm build performance, largely because this code is called multiple times
down every edge of the build graph).

This can pick up a cache hit even when it shouldn't because while the transitions
themselves properly cache, the dep rules' attributes aren't included in the cache
key computation. So applying the same transition {foo=attr.myattr} twice over
rules mydep1(myattr=a) and mydep2(myattr=b) incorrectly produces the same result,
using the faulty logic that because the Starlark code "{foo=attr.myattr}" is
the same in both cases, the output is the same. That doesn't account for the
output depending on the mutually different values of myattr.

PiperOrigin-RevId: 274869881