Make sure we do await for concurrent spawns to complete.
When a running future is cancelled, the body of the future receives an
InterruptedException that bubbles up (and that we want to use to perform
cleanup activities, like waiting for subprocesses to exit).

However, listeners attached to the future do not necessarily wait for the
InterruptedException to propagate through the body of the future... which
means our logic to await for the branch to complete was faulty.

Fix this by making the body of the branches themselves signal completion,
and only use listeners to handle the case where the branch didn't even
start running.

RELNOTES: None.
PiperOrigin-RevId: 275027140