Make --deleted_packages work for external repo
The reason `--deleted_packages` doesn't work for external repo is simply because we skip the check if the package identifier is for an external repo.

This allows us to delete specific packages from external repositories but it doesn't allow us to ignore packages by a pattern, like the `.bazelignore` file does.

Related: https://github.com/bazelbuild/bazel/issues/8854

Closes #10018.

PiperOrigin-RevId: 275033350