Wipe tree artifact directories when cancelling a local spawn.
When we enable dynamic execution without sandboxing, the local spawn is
allowed to run until the remote spawn is done executing (only when
--experimental_local_lockfree_output is given) and we are ready to
download its outputs. At that point, we cancel the local spawn and then
proceed to download the remote outputs.

This is fine in the general case but breaks with tree artifacts: the
local spawn may create extra files in the output directory during its
execution (think temporary files) that are deleted immediately by the
spawn itself before exiting. When running remotely, these files never
become part of the tree artifact manifest, but when running locally,
these files will stay behind if the local spawn is interrupted. Once
this happens, we fail tree artifact metadata validation later on because
we encounter undeclared files on disk.

To resolve this, clean up any files written by the local spawn as part
of its interrupt.

RELNOTES: None.
PiperOrigin-RevId: 275441715