Add a very simple parser for semver.
The integration with sandboxfs relies on the output of "sandboxfs --version"
to determine which version of the communication protocol to use.  As I'm
about to introduce an incompatibility in the protocol, I need the ability
to actually parse the versions.

My first attempt at this was to generalize the DottedVersion class from the
Apple rules support... but that's not desirable because that version is
intended to closely track Xcode's numbering scheme.

So, for now, add the minimal amount of code needed to parse semvers as
printed by sandboxfs.  We can add more flexibility when/if the need arises.

RELNOTES: None.
PiperOrigin-RevId: 275974959