Fix deleteTreesBelow crash if top directory goes missing.
The deleteTreesBelow method starts with a check to see if the given path
is a directory and aborts early if not.  This was masking a bug (introduced
in https://github.com/bazelbuild/bazel/commit/fac322b1ef7a9acbe71f3718143c0dae43fb5833) in the JNI code where we would not gracefully handle errors
if any syscalls failed on this top-level directory.

Fix this by removing an invalid assertion and add a race condition test
to ensure this works and that we get valid error messages.

While doing this, replace all assert()s in this algorithm with CHECK()s for
consistency with the rest of the file.

RELNOTES: None.
PiperOrigin-RevId: 276030246