Automated rollback of commit 49076cc6694e10170c2640048adbc1b1efd5c810.
*** Reason for rollback ***

The newly added tests applied to all file systems, not just the
UnixFileSystem... and they have exposed potential problems in them.
Nice! But I need to investigate what to do, so roll back in the
meantime.

RELNOTES: None.

*** Original change description ***

Fix deleteTreesBelow crash if top directory goes missing.

The deleteTreesBelow method starts with a check to see if the given path
is a directory and aborts early if not.  This was masking a bug (introduced
in https://github.com/bazelbuild/bazel/commit/fac322b1ef7a9acbe71f3718143c0dae43fb5833) in the JNI code where we would not gracefully handle errors
if any syscalls failed on this top-level directory.

Fix this by removing an invalid assertion and add a race condition test
to ensure this works and that we get valid error messages.

While doing this, replace all...

***

PiperOrigin-RevId: 276057920