Fix deleteTreesBelow crash if top directory goes missing.
This is a repeat of https://github.com/bazelbuild/bazel/commit/49076cc6694e10170c2640048adbc1b1efd5c810, which was rolled back because the new
tests added here broke when testing InMemoryFileSystem.  As it turns out,
these new tests exposed a bug of the same kind in that other file system's
delete() implementation, which I'm fixing now.

The deleteTreesBelow method starts with a check to see if the given path
is a directory and aborts early if not.  This was masking a bug (introduced
in https://github.com/bazelbuild/bazel/commit/fac322b1ef7a9acbe71f3718143c0dae43fb5833) in the JNI code where we would not gracefully handle errors
if any syscalls failed on this top-level directory.

Fix this by removing an invalid assertion and add a race condition test
to ensure this works and that we get valid error messages.

While doing this, replace all assert()s in this algorithm with CHECK()s for
consistency with the rest of the file.

RELNOTES: None.
PiperOrigin-RevId: 276072383