Fix potential race between close and write in NativeFileOutputStream.
The close() method uses a flag (closed) to track whether it has been
called and the write() method checks this flag.  The former was marked
as synchronized but the latter wasn't, which looks like a problem.

This essentially makes write() reliably throw IOException instead of
IllegalArgumentException (which comes from unix_jni.cc when write(2)
returns EBADF) when the backing file descriptor is closed.

PiperOrigin-RevId: 276349099