Replace sandboxfs --version parsing with new GnuVersionParser.
While doing this, get rid of the pre-mount version check performed "out of
band" to check if sandboxfs can be used.  Instead, do the version check
within the actual mount attempt -- and when things fail, decide whether we
have to report an error or not based on whether sandboxfs usage was
explicitly requested or selected automatically.

RELNOTES: None.
PiperOrigin-RevId: 276362964