aquery_differ no longer depends on Abseil.
This is the only Python script in the Bazel source
tree that uses abseil-py. I'm planning to remove
abseil-py from the Bazel source tree.

abseil-py is 1.8 MB, removing it will speed up Git
checkouts and shrink the distribution archive.

abseil-py is also a maintenance liability, because
we need to update it occasionally. I ran into that
just now, trying to update Bazel's Python sources
to PY3.

PiperOrigin-RevId: 276653655