Simplify the sandboxfs map/unmap interface to work in terms of "sandboxes".
Bazel creates a top-level directory within sandboxfs for each action and
makes sure that these hierarchies remain disjoint.  We'll take advantage
of this in sandboxfs to optimize the way mappings are applied.  But, for
now, make sure to propagate this information down to the sandboxfs module
so that we can later push it into sandboxfs itself.

To do this, replace the previous map/unmap primitives, which worked at
the path level, with higher-level createSandbox/destroySandbox abstractions,
which work at the level of a whole top-level subdirectory of the mount
point.

It's very likely that the RPC system of sandboxfs will adopt this same
interface, although that's still under development.

RELNOTES: None.
PiperOrigin-RevId: 276753593