Use JsonWriter instead of hand-crafting JSON.
With the upcoming sandboxfs protocol rewrite, we'll have to start accepting
JSON responses as well, which means we'll need to bring in the JSON parser.
Therefore, it doesn't hurt to also use JsonWriter to clean up the code a
little bit.

This change is a no-op.  Note that it's not optimal because we still do a
lot of in-memory string manipulation, but I don't want to worry too much
about this because this code is about to change to implement a new protocol
version.

RELNOTES: None.
PiperOrigin-RevId: 276804426