Handle packages with errors in TestsInSuiteFunction.
Clients of PackageFunction have to be careful to handle partially valid package results. Packages in an error state at least need to be flagged, so higher levels don't think the whole build succeeded. If --nokeep_going is in effect, an exception should generally be raised to abort the build.

Fixes https://github.com/bazelbuild/bazel/issues/10027.

Closes #10080.

PiperOrigin-RevId: 277227257