Guard include scanner against null grepIncludes
The grepIncludes artifact may be null if the owning rule doesn't have a
$grep_includes attribute (see CppHelper.getGrepIncludes). This seems to
be intentional. However, the remote include scanner requires that
grepIncludes is *not* null, so we need to safe-guard against it.

Unfortunately, this is brittle, because accidentally misspelling or
removing the attributes makes all grep operations fall back to the
built-in implementation.

PiperOrigin-RevId: 277499382