Automated rollback of commit 06d2f6d8ad1a7cde19c064e54e31dad7bd8cfdcf.
*** Reason for rollback ***

We must implement path case fixing on top of Skyframe (by requesting DirectoryListingValue), rather than as a JNI method. That way clients get invalidated when a path is renamed such that only the casing changes.

*** Original change description ***

Windows: implement getCorrectCasing

Implement getCorrectCasing: a function that
normalizes a path and returns the case-correct
version of it.

See https://github.com/bazelbuild/bazel/issues/8799

Closes #9435.

PiperOrigin-RevId: 277688249