bazel syntax: replace NestedSet with SkylarkNestedSet in @SkylarkCallable-annotated method results
NestedSet is not a valid Starlark value. Currently the interpreter
implicitly converts NestedSet to SkylarkNestedSet,
but it should not know about either of these types.

This change only handles the conversions in cases where the type is
explicit; there may be others where the result type is Object, for example.

In one case NestedSet<PathFragment> was replaced with String elements,
as PathFragment is also not a Starlark value. (This was actually a simplification.)

Also: add Nullable annotations to JavaCompilationInfoProvider to match reality.
PiperOrigin-RevId: 277795809