bazel skylarkbuildapi: don't return PathFragment values to Starlark
This is a user-visible change, and thus technically a breaking change,
as it causes the result of repr(x) to change from foo to "foo" for a PathFragment.
But no-one does that; in Google code, the only thing users ever to do
with a PathFragment is to pass it to str().

Soon it will be an assertion failure to return a non-Starlark value
(other than approved ones such as List and Map) from a Java annotated method.

BUG=143706853
PiperOrigin-RevId: 277811299