Add support for sandboxfs 0.2 and newer.
This change adds support for the new reconfigurations protocol in sandboxfs
0.2.0 and above without breaking compatibility with the 0.1.x series.

To do so, this splits the RealSandboxfsProcess class into two subclasses,
moving all protocol-specific logic into them.  Note that the mount flags
could also vary across versions, but as they do not yet, I'm keeping them
in the generic parent class.

As part of this change, I'm also adding unit tests for each of the
supported versions.  These are very rudimentary and do not guarantee that
Bazel can communicate with sandboxfs.  However, they provide a certain
level of safety during refactorings of the Bazel code because these tests
ensure that the protocol spoken by Bazel does not change unexpectedly.

RELNOTES: None.
PiperOrigin-RevId: 278350105